+++ AxesRangeToButton/AxesRangeToButton.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using System.Timers;
 using HidWizards.UCR.Core.Attributes;
 using HidWizards.UCR.Core.Models;
 using HidWizards.UCR.Core.Models.Binding;
@@ -7,8 +8,8 @@ using HidWizards.UCR.Core.Utilities;
 namespace AxesRangeToButton
 {
     [Plugin("Axes Range to Button", Group = "Axis", Description = "Presses a button when two axes are within a specific range")]
-    [PluginInput(DeviceBindingCategory.Range, "Axis 1")]
-    [PluginInput(DeviceBindingCategory.Range, "Axis 2")]
+    [PluginInput(DeviceBindingCategory.Range, "X Axis")]
+    [PluginInput(DeviceBindingCategory.Range, "Y Axis")]
     [PluginOutput(DeviceBindingCategory.Momentary, "Button")]
     public class AxesRangeToButton : Plugin
     {
@@ -30,11 +31,18 @@ namespace AxesRangeToButton
         [PluginGui("Y End %")]
         public double YEnd { get; set; }
 
+        [PluginGui("Tap Mode")]
+        public bool TapMode { get; set; }
+
+        [PluginGui("Tap Mode Duration")]
+        public int TapModeDuration { get; set; }
+
         private short _xStart;
         private short _xEnd;
         private short _yStart;
         private short _yEnd;
         private short _currentState;
+        private System.Timers.Timer _releaseTimer;
 
         public AxesRangeToButton()
         {
@@ -42,6 +50,7 @@ namespace AxesRangeToButton
             XEnd = 50.0;
             YStart = -50.0;
             YEnd = 50.0;
+            TapModeDuration = 50;
         }
 
         private void Initialize()
@@ -51,6 +60,8 @@ namespace AxesRangeToButton
 
             _yStart = Functions.ClampAxisRange((int)(YStart * 327.68));
             _yEnd = Functions.ClampAxisRange((int)(YEnd * 327.68));
+            _releaseTimer = new System.Timers.Timer { AutoReset = false, Interval = TapModeDuration };
+            _releaseTimer.Elapsed += DoRelease;
         }
 
         public override void Update(params short[] values)
@@ -70,8 +81,24 @@ namespace AxesRangeToButton
             }
 
             if (newState == _currentState) return;
-            WriteOutput(0, newState);
-            _currentState = newState;
+            if (TapMode)
+            {
+                if (newState == 1 && _currentState == 0)
+                {
+                    WriteOutput(0, 1);
+                    _currentState = 1;
+                    _releaseTimer.Enabled = true;
+                }
+                else if (newState == 0 && _currentState == 1)
+                {
+                    _currentState = 0;
+                }
+            }
+            else
+            {
+                WriteOutput(0, newState);
+                _currentState = newState;
+            }
         }
 
         #region Event Handling
@@ -87,6 +114,11 @@ namespace AxesRangeToButton
             Initialize();
         }
         #endregion
+
+        private void DoRelease(object sender, ElapsedEventArgs elapsedEventArgs)
+        {
+            WriteOutput(0, 0);
+        }
 
         public override PropertyValidationResult Validate(PropertyInfo propertyInfo, dynamic value)
         {