+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed 
 ### Deprecated
 ### Fixed
+- [AxesToAxesTrim Plugin] Axes no longer wrap around
 
 ##[0.8.2] - 2019-12-18
 ### Fixed